refactor: make AnswerGuide component inclusive to its logic
